field names truncating removed, added for unique constraint, keeped for relation names
